chore(remap transform): split up functions into separate modules
