+++ src/mapping/query/function/not.rs
@@ -1,0 +1,57 @@
+use super::prelude::*;
+
+#[derive(Debug)]
+pub(in crate::mapping) struct NotFn {
+    query: Box<dyn Function>,
+}
+
+impl NotFn {
+    pub(in crate::mapping) fn new(query: Box<dyn Function>) -> Self {
+        Self { query }
+    }
+}
+
+impl Function for NotFn {
+    fn execute(&self, ctx: &Event) -> Result<Value> {
+        self.query.execute(ctx).and_then(|v| match v {
+            Value::Boolean(b) => Ok(Value::Boolean(!b)),
+            _ => Err(format!("unable to perform NOT on {:?} value", v)),
+        })
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::mapping::query::path::Path;
+
+    #[test]
+    fn not() {
+        let cases = vec![
+            (
+                Event::from(""),
+                Err("path .foo not found in event".to_string()),
+                NotFn::new(Box::new(Path::from(vec![vec!["foo"]]))),
+            ),
+            (
+                Event::from(""),
+                Ok(Value::Boolean(false)),
+                NotFn::new(Box::new(Literal::from(Value::Boolean(true)))),
+            ),
+            (
+                Event::from(""),
+                Ok(Value::Boolean(true)),
+                NotFn::new(Box::new(Literal::from(Value::Boolean(false)))),
+            ),
+            (
+                Event::from(""),
+                Err("unable to perform NOT on Bytes(b\"not a bool\") value".to_string()),
+                NotFn::new(Box::new(Literal::from(Value::from("not a bool")))),
+            ),
+        ];
+
+        for (input_event, exp, query) in cases {
+            assert_eq!(query.execute(&input_event), exp);
+        }
+    }
+}
